migrate GET_PRAYER_FEATURE query into its own file
